More cleanup and relocating files
